luci-app-autodl: after installing gst-play-1.0 plug-in package, m4a and aac audio formats are supported

+++ package/network/services/luci-app-autodl/files/autodl/xmlygetlist.sh
@@ -1,9 +1,18 @@
 #!/bin/sh
 
-rm /tmp/tmp.Audioxm.list
-
 paudiopath=$(uci get autodl.@autodl[0].xmlypath)
 paudioname=$(uci get autodl.@autodl[0].xmlyname)
 
-cd /$paudiopath/$paudioname
-find *.mp3 > /tmp/tmp.Audioxm.list
+testplayer=$(opkg list-installed | grep "gst-play-1.0")
+
+if [ -e /tmp/tmp.Audioxm.list ];then
+	rm /tmp/tmp.Audioxm.list
+fi
+
+if [ ! "$testplayer" ];then
+	find $paudiopath/$paudioname/*.mp3 > /tmp/tmp.Audioxm.list
+else
+	find $paudiopath/$paudioname/*.mp3 > /tmp/tmp.Audioxm.list
+	find $paudiopath/$paudioname/*.m4a >> /tmp/tmp.Audioxm.list
+	find $paudiopath/$paudioname/*.aac >> /tmp/tmp.Audioxm.list
+fi
